working version isolated from other user's code
the .html, .js, and .css files together display a page of news 50 headlines, images, and links to original article from a fake news subreddit AND a real news subreddit (25 fake + 25 real).  it also updates a firebase database with 5 json values of each news article which will be used in the formulation of the final project, a web app/game to determine if user can tell if a news article headline is real or fake.